requirements.txt
asgiref==3.8.1
Django==5.1.4
langchain==0.1.9
langchain-mistralai==0.0.6
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
djangorestframework==3.14.0
django-cors-headers==4.3.1